requirements.txt
alembic==0.7.4
Flask-Failsafe==0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask==0.10.1
psycopg2==2.5.4
pytest==2.6.4
flake8
isort
